Split HonkCommandModule
HonkCommandModule is now defined as a partial class and I have split the individual commands into their own files.
